Adding context to arithmetic

+++ src/main/java/piquebinaries/runnable/QualityModelDeriver.java
@@ -22,6 +22,7 @@
  */
 package piquebinaries.runnable;
 
+import java.math.BigDecimal;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -105,7 +106,7 @@ public class QualityModelDeriver {
 
         // (1) Derive thresholds
         IBenchmarker benchmarker = qmDesign.getBenchmarker();
-        Map<String, Double[]> measureNameThresholdMappings = benchmarker.deriveThresholds(
+        Map<String, BigDecimal[]> measureNameThresholdMappings = benchmarker.deriveThresholds(
             benchmarkRepository, qmDesign, tools, projectRootFlag);
 
         // (2) Elicitate weights
